api/requirements.txt
googlemaps>=4.10.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0 